requirements.txt
asgiref==3.6.0
Babel==2.11.0
Django==4.1.4
django-admin-rangefilter==0.9.0
django-extensions==3.2.1
django-filter==22.1
django-money==3.0.0
djangorestframework==3.14.0
py-moneyed==2.0
python-slugify==8.0.0
pytz==2022.7.1
sqlparse==0.4.3
text-unidecode==1.3
typing_extensions==4.5.0
